Mod: redo eval from mceval -> eceval
Even though the question says implement the Explicit Control Evaluator,
I based Eval on the Metacircular Evaluator because that's easier/tidier
and I didn't understand why it should be based on eceval insted of the
mceval.  The answer is demonstrated by Ex 1.05, this should 'hang', i.e,
it should result in an infinite loop.  But instead it crashed becauses
of a stackoverflow, although C does support some tail call optimization
it doesn't appear (and can't be guaranteed) to consistently do tail
call eliminiation esp. with the complex arrangement of calls we have
with eval/apply.